:truck: Standardize page naming; suffix with "Page"
